requirements.txt
discord.py
discord.py[voice]
youtube_dl